+++ common/src/main/kotlin/flank.common/Environment.kt
@@ -1,0 +1,3 @@
+package flank.common
+
+fun isCI() = System.getenv("CI") != null
