+++ flank-scripts/.gitignore
@@ -24,4 +24,3 @@ out/
 buildSrc/.gradle/
 buildSrc/build
 local.properties
-flank-scripts.properties
