+++ flank-scripts/README.md
@@ -308,7 +308,7 @@ Applies a pre-commit hook that runs Ktlint on a commit and fails if there are li
 Applies Ktlint styling to the current idea project, so that correct code styling is upheld.
 
 ## Testing
-To test your script with different settings use `flank-scripts.properties` file. Uncomment and replace with desired values.
+To test your script with different settings use `flank-debug.properties` file. Uncomment and replace with desired values.
 Properties are skipped by git and should not be attached to a commit. Note, `test` task ignores your own properties and will use the default.
 
 #### List of possible configs
