--- flank-scripts/flank-scripts.properties
@@ -1,11 +1,0 @@
-#zenhub.repo-id=84221974
-#
-#repo.flank=Flank/flank
-#repo.gcloud_cli=Flank/gcloud_cli
-#repo.test-artifacts=Flank/test_artifacts
-#
-#integration.workflow-filename=full_suite_integration_tests.yml
-#integration.issue-poster=github-actions[bot]
-#
-#sdk-check.workflow-filename=update_dependencies_and_client.yml
-#sdk-check.issue-poster=github-actions[bot]
