+++ flank-scripts/src/main/kotlin/flank/scripts/shell/ios/SetupIosEnvCommand.kt
@@ -1,9 +1,9 @@
 package flank.scripts.shell.ios
 
 import com.github.ajalt.clikt.core.CliktCommand
+import flank.common.iOSTestProjectsPath
 import flank.scripts.shell.ops.EARL_GREY_EXAMPLE
 import flank.scripts.shell.utils.failIfWindows
-import flank.scripts.shell.utils.iOSTestProjectsPath
 import flank.scripts.utils.downloadCocoaPodsIfNeeded
 import flank.scripts.utils.installPodsIfNeeded
 import java.nio.file.Paths
