+++ flank-scripts/src/main/kotlin/flank/scripts/shell/ios/UniversalFramework.kt
@@ -1,7 +1,7 @@
 package flank.scripts.shell.ios
 
 import com.github.ajalt.clikt.core.CliktCommand
-import flank.scripts.shell.utils.currentPath
+import flank.common.currentPath
 import flank.scripts.shell.utils.failIfWindows
 import flank.scripts.utils.runCommand
 import java.nio.file.Paths
