+++ flank-scripts/src/main/kotlin/flank/scripts/shell/ops/BuildIosIPA.kt
@@ -1,8 +1,8 @@
 package flank.scripts.shell.ops
 
+import flank.common.flankFixturesIosTmpPath
 import flank.scripts.shell.ios.createXcodeArchiveCommand
 import flank.scripts.shell.ios.createXcodeExportArchiveCommand
-import flank.scripts.shell.utils.flankFixturesIosTmpPath
 import flank.scripts.shell.utils.pipe
 import flank.scripts.utils.downloadCocoaPodsIfNeeded
 import flank.scripts.utils.downloadXcPrettyIfNeeded
